+++ .gitignore
@@ -1,0 +1,7 @@
+.DS_Store
+.idea
+*.log
+tmp/
+
+node_modules/
+*.screenflow
