+++ originalExample/docs/01_query_in_graphiql.md
