+++ originalExample/docs/03_authed_data_in_graphiql.md
