+++ originalExample/public/favicon.ico
