+++ originalExample/public/logo192.png
