+++ originalExample/public/logo512.png
