+++ originalExample/public/manifest.json
