+++ originalExample/src/index.js
