+++ originalExample/src/utils/auth.js
